[Lite Interpreter] Put mobile-used prim operators in a new file
Differential Revision: [D20830606](https://our.internmc.facebook.com/intern/diff/D20830606/)

**NOTE FOR REVIEWERS**: This PR has internal Facebook specific changes or comments, please review them on [Phabricator](https://our.internmc.facebook.com/intern/diff/D20830606/)!

[ghstack-poisoned]
